+++ project.greenfoot
@@ -1,28 +1,39 @@
 #Greenfoot project file
-class.MyWorld.image=MAP\ FINAL.png
+class.ButtonPlay.image=01_Red_Buttons_Pixel.png
+class.Homepage.image=Homepage.png
+class.PlayGround.image=MAP\ FINAL.png
 class.Player.image=00_idle_down.png
-dependency1.from=MyWorld
-dependency1.to=Player
+dependency1.from=Homepage
+dependency1.to=ButtonPlay
 dependency1.type=UsesDependency
-dependency2.from=MyWorld
-dependency2.to=Enemy
+dependency2.from=PlayGround
+dependency2.to=Player
 dependency2.type=UsesDependency
-dependency3.from=Enemy
-dependency3.to=Player
+dependency3.from=PlayGround
+dependency3.to=Weapon
 dependency3.type=UsesDependency
-dependency4.from=Weapon
-dependency4.to=Player
+dependency4.from=PlayGround
+dependency4.to=Enemy
 dependency4.type=UsesDependency
-dependency5.from=Weapon
-dependency5.to=Projectile
+dependency5.from=Enemy
+dependency5.to=Player
 dependency5.type=UsesDependency
+dependency6.from=ButtonPlay
+dependency6.to=PlayGround
+dependency6.type=UsesDependency
+dependency7.from=Weapon
+dependency7.to=Player
+dependency7.type=UsesDependency
+dependency8.from=Weapon
+dependency8.to=Projectile
+dependency8.type=UsesDependency
 editor.fx.0.height=737
 editor.fx.0.width=814
-editor.fx.0.x=360
-editor.fx.0.y=58
-height=734
-package.numDependencies=5
-package.numTargets=6
+editor.fx.0.x=504
+editor.fx.0.y=284
+height=926
+package.numDependencies=8
+package.numTargets=9
 project.charset=UTF-8
 publish.hasSource=false
 publish.locked=true
@@ -38,49 +49,70 @@ readme.x=10
 readme.y=10
 simulation.speed=50
 target1.height=70
-target1.name=Player
+target1.name=Homepage
 target1.showInterface=false
 target1.type=ClassTarget
 target1.width=120
 target1.x=0
 target1.y=0
 target2.height=70
-target2.name=Char
+target2.name=Player
 target2.showInterface=false
 target2.type=ClassTarget
 target2.width=120
 target2.x=0
 target2.y=0
 target3.height=70
-target3.name=Enemy
+target3.name=PlayGround
 target3.showInterface=false
 target3.type=ClassTarget
 target3.width=120
 target3.x=0
 target3.y=0
 target4.height=70
-target4.name=MyWorld
+target4.name=Char
 target4.showInterface=false
 target4.type=ClassTarget
 target4.width=120
 target4.x=0
 target4.y=0
 target5.height=70
-target5.name=Weapon
+target5.name=Enemy
 target5.showInterface=false
 target5.type=ClassTarget
 target5.width=120
 target5.x=0
 target5.y=0
 target6.height=70
-target6.name=Projectile
+target6.name=ButtonPlay
 target6.showInterface=false
 target6.type=ClassTarget
 target6.width=120
 target6.x=0
 target6.y=0
+target7.height=70
+target7.name=MyWorld
+target7.showInterface=false
+target7.type=ClassTarget
+target7.width=120
+target7.x=0
+target7.y=0
+target8.height=70
+target8.name=Weapon
+target8.showInterface=false
+target8.type=ClassTarget
+target8.width=120
+target8.x=0
+target8.y=0
+target9.height=70
+target9.name=Projectile
+target9.showInterface=false
+target9.type=ClassTarget
+target9.width=120
+target9.x=0
+target9.y=0
 version=3.1.0
-width=1294
-world.lastInstantiated=MyWorld
+width=1550
+world.lastInstantiated=Homepage
 xPosition=0
 yPosition=0
